Refactored codes in different files for navbar
